+++ spec/shelly/app_spec.rb
@@ -479,8 +479,8 @@ describe Shelly::App do
       @app.stub(:attributes => {"system_user" => "system_user"})
       @client.stub(:tunnel).and_return(
         {"host" => "console.example.com", "port" => "40010", "user" => "foo"})
-      @app.should_receive(:system).with("rsync -avz -e 'ssh -o StrictHostKeyChecking=no -p 40010 -l foo' --progress /path console.example.com:/home/system_user/disk")
-      @app.upload("/path")
+      @app.should_receive(:system).with("rsync --archive --verbose --compress --relative -e 'ssh -o StrictHostKeyChecking=no -p 40010 -l foo' --progress /path console.example.com:/home/system_user/disk/.")
+      @app.upload("/path", ".")
     end
   end
 
@@ -489,7 +489,7 @@ describe Shelly::App do
       @app.stub(:attributes => {"system_user" => "system_user"})
       @client.stub(:tunnel).and_return(
         {"host" => "console.example.com", "port" => "40010", "user" => "foo"})
-      @app.should_receive(:system).with("rsync -avz -e 'ssh -o StrictHostKeyChecking=no -p 40010 -l foo' --progress console.example.com:/home/system_user/disk/. /tmp")
+      @app.should_receive(:system).with("rsync --archive --verbose --compress  -e 'ssh -o StrictHostKeyChecking=no -p 40010 -l foo' --progress console.example.com:/home/system_user/disk/. /tmp")
       @app.download(".", "/tmp")
     end
   end